requirement.txt
aiosmtplib==2.0.2
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.6
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.2
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
Jinja2==3.1.3
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.2.1
pygments==2.17.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
rsa==4.9
sendgrid==6.11.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starkbank-ecdsa==2.2.0
starlette==0.37.2
typer==0.12.3
typing-extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zipp==3.18.1
